% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{find.breaks}
\alias{find.breaks}
\title{Estimate crossover locations}
\usage{
find.breaks(cross, chr = NULL)
}
\arguments{
\item{cross}{An object of class \code{cross}. (This must be a backcross,
RIL, or intercross.) See \code{\link[qtl:read.cross]{qtl::read.cross()}} for details.}

\item{chr}{Optional set of chromosomes on which to look for crossovers.  If
NULL, all chromosomes are considered.}
}
\value{
If only one chromosome is considered, this is a list with one
component for each individual.  If multiple chromosomes were considered,
this is a list with one element for each chromosome, each of which is a list
with one element for each individual, as above.

For backcrosses and RIL, the componenets for the individuals are
\code{numeric(0)} if there were no crossovers or a vector giving the
crossover locations.  The length of the chromosome (in cM) is saved as an
attribute.  (Note that the format is the same as the output of
\code{\link[=simStahl]{simStahl()}}.)

For an intercross, the components for the individuals are themselves lists
with all possible allocations of the crossovers to the two meiotic products;
each component of this list is itself a list with two components,
corresponding to the two meiotic products.
}
\description{
Estimate the locations of crossovers in a backcross.
}
\details{
This works only a backcross, RIL, or intercross.  We use the function
\code{\link[qtl:locateXO]{qtl::locateXO()}} in R/qtl.  Crossovers are estimated to be at the
midpoint of the interval between the nearest flanking typed markers.
}
\examples{

data(bssbsb)

# crossover locations on chromosome 1
xoloc1 <- find.breaks(bssbsb, chr=1)

# crossover locations on all chromosomes
xoloc <- find.breaks(bssbsb)

}
\seealso{
\code{\link[=convertxoloc]{convertxoloc()}}, \code{\link[=fitGamma]{fitGamma()}},
\code{\link[=simStahl]{simStahl()}}
}
\author{
Karl W Broman, \email{broman@wisc.edu}
}
\keyword{utilities}
