% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{kfunc}
\alias{kfunc}
\title{estimate Ripley's K function}
\usage{
kfunc(x, d = seq(0, 100, by = 0.1), lengths, exclude = 0, tol = 0.000001)
}
\arguments{
\item{x}{list with sorted locations of the data}

\item{d}{values at which to calculate the function}

\item{lengths}{lengths of segments studied}

\item{exclude}{distance to exclude}

\item{tol}{tolerance value}
}
\value{
data frame with \code{d}, \code{k}, and \code{se}
}
\description{
estimate the 1-d version of Ripley's K function
}
\seealso{
\code{\link{gammacoi}}, \code{\link{stahlcoi}}, \code{\link{coincidence}}
}
\keyword{models}

