% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{recrate2scanone}
\alias{recrate2scanone}
\title{Convert recrate to scanone format}
\usage{
recrate2scanone(recrate, phymap)
}
\arguments{
\item{recrate}{A list of results from \code{\link{est.recrate}}}

\item{phymap}{A list of vectors of Mbp positions of markers}
}
\value{
A data frame with class \code{"scanone"}, in the format output by \code{\link[qtl]{scanone}}.
}
\description{
Convert the result of \code{\link{est.recrate}} to the format
output by R/qtl's \code{\link[qtl]{scanone}} function.
}
\author{
Karl W Broman, \email{kbroman@biostat.wisc.edu}
}
\seealso{
\code{\link{est.recrate}}
}
\keyword{models}

