\name{xoi-internal}
\alias{kfunc}
\alias{recrate2scanone}

\title{Internal xoi functions}

\description{
  Internal xoi functions. These are generally not to be called by the
  user.
}

\usage{
kfunc(x, d=seq(0,100,by=0.1), lengths, exclude=0, tol=1e-6)
recrate2scanone(recrate, phymap)
}

\author{Karl W Broman, \email{kbroman@biostat.wisc.edu} }

\keyword{internal}
