% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xoi-package.R
\docType{package}
\name{xoi-package}
\alias{xoi}
\alias{xoi-package}
\title{xoi: Tools for Analyzing Crossover Interference}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Analysis of crossover interference in experimental crosses, particularly regarding the gamma model. See, for example, Broman and Weber (2000) \doi{10.1086/302923}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kbroman/xoi}
  \item Report bugs at \url{https://github.com/kbroman/xoi/issues}
}

}
\author{
\strong{Maintainer}: Karl W Broman \email{broman@wisc.edu} (\href{https://orcid.org/0000-0002-4914-6671}{ORCID})

Authors:
\itemize{
  \item Karl W Broman \email{broman@wisc.edu} (\href{https://orcid.org/0000-0002-4914-6671}{ORCID})
}

Other contributors:
\itemize{
  \item Il-Youp Kwak (\href{https://orcid.org/0000-0002-7117-7669}{ORCID}) [contributor]
}

}
\keyword{internal}
