\name{xoiversion}
\alias{xoiversion}

\title{Installed version of R/xoi}

\description{
  Print the version number of the currently installed version of R/xoi.
}

\usage{xoiversion()}

%\arguments{
%}

\value{
  A character string with the version number of the currently installed
  version of R/xoi.
}

\author{Karl W Broman, \email{kbroman@biostat.wisc.edu} }

\examples{
  xoiversion()
}


\keyword{print}
