% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/xoiversion.R
\name{xoiversion}
\alias{xoiversion}
\title{Installed version of R/xoi}
\usage{
xoiversion()
}
\value{
A character string with the version number of the currently
installed version of R/xoi.
}
\description{
Print the version number of the currently installed version of R/xoi.
}
\examples{
xoiversion()
}
\author{
Karl W Broman, \email{kbroman@biostat.wisc.edu}
}
\keyword{print}

